requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.8
Django==5.0.6
django-anymail==11.0.1
django-cors-headers==4.4.0
django-ipware==7.0.1
django-redis==5.4.0
django-rest-authemail==2.1.7
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
iniconfig==2.0.0
kombu==5.4.2
nose==1.3.7
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
redis==5.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
